+++ ContentPatcher/release-notes.md
@@ -4,6 +4,8 @@
 ## Upcoming release
 * Added 'pinned key' feature to the `Random` token, which significantly expands what you can do with it (see readme).
 * Added token support in `FromArea` and `ToArea` fields (thanks to spacechase0!).
+* Added new tokens:
+  * `Lowercase` and `Uppercase` let you transform input text.
 * `Random` token is now bounded if the choices are immutable.
 * Internal optimizations for asset loading/editing.
 * Fixed incorrect warning when an `EditData` patch references a file that doesn't exist when it should be skipped due to a `HasFile` condition.